chore: fix tests
Just combining the backpressure tests and the math tests. There were issues
with the tests because of port collisions.
